Began working on Go version.
This is mostly just to play with go, but it DOES make a lot more sense
to do the checks concurrently rather than sequentially.

So far, began working on the general structure. There is a simple
executable for testing purposes (updatechecks.go), this will eventually
do more.

Utils.go contains utility functions like getting the current version in
AUR, and checking all programs for out-of-date-ness. programs.go
contains the code to actually check the version for each program.